Deixando o mundo mais seguro

+++ Problemas/decrypt.sh
@@ -1,2 +1,4 @@
 aescrypt -d -k ~/Documents/secret.key $1
-rm $1
+if [ "$?" = "0" ]; then
+	rm $1
+fi
